Versao 1.5.4

+++ src/main/java/com/fabriciosanches/fichatecnica/dtos/ItemProdutoDTO.java
@@ -3,7 +3,7 @@ package com.fabriciosanches.fichatecnica.dtos;
 
 import java.math.BigDecimal;
 
-public record ItemProdutoDTO(Long cdItem, Long cdProduto, Integer qtdItem,
+public record ItemProdutoDTO(Long cdItem, Long cdProduto, Double qtdItem,
               Long cdUnidadeMedida, BigDecimal vlrItem) {
