+++ src/main/java/com/fabriciosanches/fichatecnica/dtos/ProdutoCompletoDTO.java
@@ -2,7 +2,7 @@ package com.fabriciosanches.fichatecnica.dtos;
 
 import java.math.BigDecimal;
 
-public record ProdutoCompletoDTO(String nomeProduto, String nomeItem, Integer qtdeItem,
+public record ProdutoCompletoDTO(String nomeProduto, String nomeItem, Double qtdeItem,
                                  Long cdUnidade,
                                  BigDecimal valorItem) {
